squash ref: move pairing code analytics to common/analytics
