+++ spot-client/src/common/analytics/analyticsEvents.js
